+++ client/layout/test/index.js
@@ -18,6 +18,12 @@ describe( 'index', () => {
 									items: {},
 									lastTimeShown: {},
 								},
+								purchases: {
+									hasLoadedUserPurchasesFromServer: false,
+								},
+								sites: {
+									items: {},
+								},
 							} ),
 							subscribe: () => {},
 						} }
