+++ client/lib/signup/flow-controller.ts
@@ -16,7 +16,8 @@ import {
 	reject,
 	reduce,
 } from 'lodash';
-import { Store, Unsubscribe as ReduxUnsubscribe } from 'redux';
+import { Store, Unsubscribe as ReduxUnsubscribe, AnyAction } from 'redux';
+import { reloadProxy, requestAllBlogsAccess } from 'wpcom-proxy-request';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
 import { logToLogstash } from 'calypso/lib/logstash';
 import wpcom from 'calypso/lib/wp';
@@ -356,6 +357,11 @@ export default class SignupFlowController {
 
 		if ( dependencies.bearer_token && ! wpcom.isTokenLoaded() ) {
 			wpcom.loadToken( dependencies.bearer_token );
+			// Reload proxy to refresh auth status for parts that incorrectly use wpcom-proxy-request in Calypso.
+			reloadProxy();
+			requestAllBlogsAccess();
+			// Fetch current user to refresh auth status and re-render all dependent components.
+			this._reduxStore.dispatch( fetchCurrentUser() as unknown as AnyAction );
 		}
 
 		for ( const pendingStep of pendingSteps ) {
